% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{contour_read}
\alias{contour_read}
\title{contour_read}
\usage{
contour_read(pdf_path)
}
\arguments{
\item{pdf_path}{file.path to input pdf}
}
\description{
Read a pdf into a lines object
}
\examples{
\dontrun{
pdf <- system.file("extdata/1835300a.pdf", package = "contour2bathy")
res <- contour_read(pdf_path = pdf)
par(mfrow = c(1,1))
sp::plot(res)

}
}

