% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{contour_read}
\alias{contour_read}
\title{contour_read}
\usage{
contour_read(pdf_path, smallest_length = 1e-05, grow_radius = 2,
  clean_thresh = 0.08)
}
\arguments{
\item{pdf_path}{file.path to input pdf}

\item{smallest_length}{numeric threshold defining the smallest line-length to be returned}

\item{grow_radius}{numeric value defining how much to "grow" lines prior to thinning}

\item{clean_thresh}{numeric threshold defining the smallest allowable "dangle"}
}
\description{
Read a pdf into a lines object
}
\examples{
\dontrun{
pdf <- system.file("extdata/1835300a.pdf", package = "contour2bathy")
res <- contour_read(pdf_path = pdf)
par(mfrow = c(1,1))
sp::plot(res)

}
}
